requirements_raw.txt
django_extensions
jupyter
django-debug-toolbar
psycopg2-binary 
django
djangorestframework
djangorestframework-simplejwt
django-crispy-forms
dj-database-url
django-multiselectfield
python-decouple
autopep8 
django-cors-headers
Pillow
crispy-tailwind
django-phonenumber-field
phonenumbers
django-phone-verify

django-simple-history
django-model-utils
django-simpleui
django-material-admin
whitenoise
django-tenant-schemas